# refactor: migrate test runner from Mocha to Vitest