--- a/packages/@textlint/ast-tester/src/index.ts
+++ b/packages/@textlint/ast-tester/src/index.ts
@@ -1,8 +1,8 @@
 // LICENSE : MIT
 "use strict";
-import * as assert from "assert";
+import * as assert from "node:assert";
 import { TxtNode } from "@textlint/ast-node-types";
-import { test as UnistTest } from "./unist-test";
+import { test as UnistTest } from "./unist-test.js";
 import debug0 from "debug";
 
 const debug = debug0("textlint/ast-tester");