--- a/packages/@textlint/kernel/.gitignore
+++ b/packages/@textlint/kernel/.gitignore
@@ -2,3 +2,5 @@
 out/
 lib/
 module/
+
+coverage