--- a/packages/@textlint/kernel/src/descriptor/TextlintLintableRuleDescriptor.ts
+++ b/packages/@textlint/kernel/src/descriptor/TextlintLintableRuleDescriptor.ts
@@ -1,7 +1,7 @@
 // LICENSE : MIT
 "use strict";
-import { TextlintKernelRule } from "../textlint-kernel-interface";
-import { assertRuleShape, getLinter } from "./rule-creator-helper";
+import { TextlintKernelRule } from "../textlint-kernel-interface.js";
+import { assertRuleShape, getLinter } from "./rule-creator-helper.js";
 import type { TextlintRuleModule, TextlintRuleOptions, TextlintRuleReporter } from "@textlint/types";
 import { deepEqual } from "fast-equals";