--- a/packages/@textlint/kernel/src/descriptor/index.ts
+++ b/packages/@textlint/kernel/src/descriptor/index.ts
@@ -1,8 +1,8 @@
-export { TextlintKernelDescriptor } from "./TextlintKernelDescriptor";
-export { TextlintRuleDescriptors } from "./TextlintRuleDescriptors";
-export { TextlintLintableRuleDescriptor } from "./TextlintLintableRuleDescriptor";
-export { TextlintFixableRuleDescriptor } from "./TextlintFixableRuleDescriptor";
-export { TextlintFilterRuleDescriptors } from "./TextlintFilterRuleDescriptors";
-export { TextlintFilterRuleDescriptor } from "./TextlintFilterRuleDescriptor";
-export { TextlintPluginDescriptors } from "./TextlintPluginDescriptors";
-export { TextlintPluginDescriptor } from "./TextlintPluginDescriptor";
+export { TextlintKernelDescriptor } from "./TextlintKernelDescriptor.js";
+export { TextlintRuleDescriptors } from "./TextlintRuleDescriptors.js";
+export { TextlintLintableRuleDescriptor } from "./TextlintLintableRuleDescriptor.js";
+export { TextlintFixableRuleDescriptor } from "./TextlintFixableRuleDescriptor.js";
+export { TextlintFilterRuleDescriptors } from "./TextlintFilterRuleDescriptors.js";
+export { TextlintFilterRuleDescriptor } from "./TextlintFilterRuleDescriptor.js";
+export { TextlintPluginDescriptors } from "./TextlintPluginDescriptors.js";
+export { TextlintPluginDescriptor } from "./TextlintPluginDescriptor.js";