--- a/packages/@textlint/kernel/src/linter/linter-processor.ts
+++ b/packages/@textlint/kernel/src/linter/linter-processor.ts
@@ -1,12 +1,12 @@
 // LICENSE : MIT
 "use strict";
-import LinterTask from "../task/linter-task";
-import TaskRunner from "../task/task-runner";
-import { TextlintKernelConstructorOptions } from "../textlint-kernel-interface";
+import LinterTask from "../task/linter-task.js";
+import TaskRunner from "../task/task-runner.js";
+import { TextlintKernelConstructorOptions } from "../textlint-kernel-interface.js";
 
-import MessageProcessManager from "../messages/MessageProcessManager";
-import { TextlintFilterRuleDescriptors, TextlintRuleDescriptors } from "../descriptor";
-import { invariant } from "../util/invariant";
+import MessageProcessManager from "../messages/MessageProcessManager.js";
+import { TextlintFilterRuleDescriptors, TextlintRuleDescriptors } from "../descriptor/index.js";
+import { invariant } from "../util/invariant.js";
 import type { TextlintSourceCode, TextlintPluginProcessor, TextlintResult } from "@textlint/types";
 
 export interface LinterProcessorArgs {