--- a/packages/@textlint/kernel/src/messages/filter-ignored-process.ts
+++ b/packages/@textlint/kernel/src/messages/filter-ignored-process.ts
@@ -1,7 +1,7 @@
 // LICENSE : MIT
 "use strict";
-import MessageType from "../shared/type/MessageType";
-import { IgnoreReportedMessage, LintReportedMessage } from "../task/textlint-core-task";
+import MessageType from "../shared/type/MessageType.js";
+import { IgnoreReportedMessage, LintReportedMessage } from "../task/textlint-core-task.js";
 
 /**
  * the `index` is in the `range` and return true.