--- a/packages/@textlint/kernel/src/task/fixer-task.ts
+++ b/packages/@textlint/kernel/src/task/fixer-task.ts
@@ -1,12 +1,12 @@
 // LICENSE : MIT
 "use strict";
-import CoreTask from "./textlint-core-task";
-import { TextlintKernelConstructorOptions } from "../textlint-kernel-interface";
-import { TextlintFilterRuleDescriptors, TextlintFixableRuleDescriptor } from "../descriptor";
+import CoreTask from "./textlint-core-task.js";
+import { TextlintKernelConstructorOptions } from "../textlint-kernel-interface.js";
+import { TextlintFilterRuleDescriptors, TextlintFixableRuleDescriptor } from "../descriptor/index.js";
 import type { TextlintSourceCode } from "@textlint/types";
-import { getSeverity } from "../shared/rule-severity";
-import { TextlintFilterRuleContextImpl } from "../context/TextlintFilterRuleContextImpl";
-import { TextlintRuleContextImpl } from "../context/TextlintRuleContextImpl";
+import { getSeverity } from "../shared/rule-severity.js";
+import { TextlintFilterRuleContextImpl } from "../context/TextlintFilterRuleContextImpl.js";
+import { TextlintRuleContextImpl } from "../context/TextlintRuleContextImpl.js";
 import _debug from "debug";
 
 const debug = _debug("textlint:TextLintCoreTask");