--- a/packages/@textlint/kernel/src/task/linter-task.ts
+++ b/packages/@textlint/kernel/src/task/linter-task.ts
@@ -1,12 +1,12 @@
 // LICENSE : MIT
 "use strict";
-import CoreTask from "./textlint-core-task";
-import { TextlintKernelConstructorOptions } from "../textlint-kernel-interface";
-import { TextlintFilterRuleDescriptors, TextlintRuleDescriptors } from "../descriptor";
+import CoreTask from "./textlint-core-task.js";
+import { TextlintKernelConstructorOptions } from "../textlint-kernel-interface.js";
+import { TextlintFilterRuleDescriptors, TextlintRuleDescriptors } from "../descriptor/index.js";
 import type { TextlintSourceCode } from "@textlint/types";
-import { getSeverity } from "../shared/rule-severity";
-import { TextlintRuleContextImpl } from "../context/TextlintRuleContextImpl";
-import { TextlintFilterRuleContextImpl } from "../context/TextlintFilterRuleContextImpl";
+import { getSeverity } from "../shared/rule-severity.js";
+import { TextlintRuleContextImpl } from "../context/TextlintRuleContextImpl.js";
+import { TextlintFilterRuleContextImpl } from "../context/TextlintFilterRuleContextImpl.js";
 
 export interface TextLintCoreTaskArgs {
     config: TextlintKernelConstructorOptions;