--- a/packages/@textlint/kernel/src/task/task-runner.ts
+++ b/packages/@textlint/kernel/src/task/task-runner.ts
@@ -4,7 +4,7 @@ import CoreTask, {
     default as TextLintCoreTask,
     IgnoreReportedMessage,
     LintReportedMessage
-} from "./textlint-core-task";
+} from "./textlint-core-task.js";
 
 export default class TaskRunner {
     /**