--- a/packages/@textlint/kernel/src/util/parse-by-plugin.ts
+++ b/packages/@textlint/kernel/src/util/parse-by-plugin.ts
@@ -1,4 +1,4 @@
-import { isPluginParsedObject } from "./isPluginParsedObject";
+import { isPluginParsedObject } from "./isPluginParsedObject.js";
 import type { TextlintPluginProcessor } from "@textlint/types";
 import { TxtNode } from "@textlint/ast-node-types";