--- a/packages/@textlint/kernel/src/util/timing.ts
+++ b/packages/@textlint/kernel/src/util/timing.ts
@@ -4,7 +4,7 @@
  * @copyright 2014 Brandon Mills. All rights reserved.
  */
 "use strict";
-import Logger from "./logger";
+import Logger from "./logger.js";
 
 /* istanbul ignore next */
 /**