--- a/packages/@textlint/linter-formatter/src/formatters/table.ts
+++ b/packages/@textlint/linter-formatter/src/formatters/table.ts
@@ -6,7 +6,7 @@
 
 "use strict";
 import type { TextlintMessage } from "@textlint/types";
-import { FormatterOptions } from "../FormatterOptions";
+import { FormatterOptions } from "../FormatterOptions.js";
 
 //------------------------------------------------------------------------------
 // Requirements