## Overview

This PR migrates the test runner from Mocha to Vitest across the entire textlint monorepo.

## Changes

### Test Runner Migration
- **Replaced Mocha with Vitest** as the primary test runner
- **Added Vitest configuration** (`vitest.config.mts`) with optimized settings for the monorepo
- **Removed all `.mocharc.json` files** from individual packages
- **Updated package.json scripts** to use `vitest run` and `vitest watch` commands

### Dependencies
- **Added `vitest` dependency** to all packages that previously used Mocha
- **Removed Mocha-related dependencies**: `mocha`, `@types/mocha`
- **Updated root package.json** to include Vitest as a dev dependency

### Test Files
- **Updated import statements** in test files:
  - Changed from `import * as assert from "assert"` to `import * as assert from "node:assert"`
  - Added `import { describe, it } from "vitest"` for test functions
  - Updated module imports to use `.js` extensions for better ESM compatibility

### Configuration
- **Vitest configuration** includes:
  - Node.js environment for server-side testing
  - Custom test timeout (10 seconds)
  - Bail on first failure for faster feedback
  - Support for TypeScript and JSON files
  - Pass with no tests to handle packages without tests

### pnpm Workspace
- **Updated pnpm-workspace.yaml** to include esbuild in `onlyBuiltDependencies` for better build performance

## Packages Maintaining Mocha

The following packages continue to use Mocha as their test runner:

- **`packages/textlint-scripts`**: Maintains Mocha dependency and test scripts
- **`packages/textlint-tester`**: Maintains Mocha dependency and test scripts

**Reason**: `textlint-tester` has a direct dependency on Mocha's global `describe` and `it` functions. The package exports test utilities that rely on these functions being available in the global scope. Migrating these packages would require significant API changes and could break existing user code that depends on the Mocha integration.

**Future Plan**: We plan to migrate to Node.js's built-in `node --test` in the future, which would provide a path to remove the Mocha dependency while maintaining compatibility.

## Benefits

1. **Faster Test Execution**: Vitest provides faster test execution with better parallelization
2. **Better ESM Support**: Improved ES module support compared to Mocha
3. **Modern Testing Features**: Access to modern testing capabilities and better developer experience
4. **Consistency**: Unified test runner across the entire monorepo
5. **Better TypeScript Integration**: Native TypeScript support without additional configuration

## CI Performance Comparison

The migration from Mocha to Vitest shows significant performance improvements in CI execution:

### Before (Mocha) - [Run #15652161176](https://github.com/textlint/textlint/actions/runs/15652161176)
- **Total Duration**: 9m 34s
- **Status**: ✅ Success
- **Jobs**: All test jobs completed successfully
- **Test Framework**: Mocha with individual `.mocharc.json` configurations

### After (Vitest) - [Run #15660021249](https://github.com/textlint/textlint/actions/runs/15660021249?pr=1544)
- **Total Duration**: 6m 32s ⚡
- **Status**: ✅ Success  
- **Jobs**: All test jobs completed successfully
- **Test Framework**: Vitest with unified configuration

### Performance Improvement
- **31.6% faster CI execution** (9m 34s → 6m 32s)
- **3m 2s reduction** in total CI time
- **Consistent performance** across all Node.js versions (20, 22) and platforms (Ubuntu, Windows)

The significant CI performance improvement demonstrates Vitest's efficiency in handling large monorepo test suites with better parallelization and optimized execution.

## Compatibility

- All existing tests continue to work with minimal changes
- Test syntax remains largely the same (using `describe` and `it`)
- API compatibility maintained for existing test assertions

## Files Changed

- **350 files** modified across the monorepo
- **+13,720 additions, -12,415 deletions** reflecting the comprehensive migration
- All packages under `packages/@textlint/*` and `examples/*` updated

## Testing

All existing tests pass with the new Vitest runner, ensuring no regression in functionality.